Move loadPage to EC2Store
