+++ CloudyCider/models/EC2Store.swift
@@ -8,6 +8,8 @@
 
 import Combine
 import SwiftUI
+import EC2
+import AWSSDKSwiftCore
 
 final class EC2Store: BindableObject {
   let didChange = PassthroughSubject<(), Never>()
@@ -19,4 +21,54 @@ final class EC2Store: BindableObject {
       }
     }
   }
+    
+    func loadPage() {
+        // MARK: - EC2Init
+        
+        let accessKeyId = ProcessInfo.processInfo.environment["accessKeyId"]
+        let secretAccessKey = ProcessInfo.processInfo.environment["secretAccessKey"]
+        let ec2 = EC2(
+            accessKeyId: accessKeyId,
+            secretAccessKey: secretAccessKey,
+            region: .euwest2
+        )
+        let instancesRequest = EC2.DescribeInstancesRequest()
+        do {
+            try ec2.describeInstances(instancesRequest).whenComplete { response in
+                do {
+                    guard let reservations = try response.get().reservations else { return }
+                    
+                    // create temp storage for the instances list
+                    var nextStoreInstances: [EC2Instance] = []
+                    
+                    for reservation in reservations {
+                        guard let instances = reservation.instances else { return }
+                        
+                        for instance in instances {
+                            // get the instance state and the instance name
+                            guard let state = instance.state,
+                                let statusName = state.name,
+                                let tags = instance.tags,
+                                let nameTag = tags.filter({ $0.key == "Name" }).first,
+                                let name = nameTag.value else { return }
+                            
+                            // add new instance information to the list
+                            let instanceInformation = EC2Instance(name: name, status: statusName)
+                            nextStoreInstances.append(instanceInformation)
+                        }
+                    }
+                    // replace old instances stored in store with with the new one fetched from aws
+                    self.result = .success(nextStoreInstances)
+                } catch let error as AWSError {
+                    self.result = .failure(error)
+                } catch {
+                    print(error)
+                    self.result = .failure(error)
+                }
+            }
+        } catch {
+            print(error)
+            result = .failure(error)
+        }
+    }
 }
